==> 1-API-data-extraction-1.ipynb <==
# Necessary libraries
import requests
import pandas as pd

# Set your API key
api_key = "24efa7fb0e6e8e80b521506cdfcd1f76"

# Base API URL
base_url = "https://manifesto-project.wzb.eu/api/v1/"

#get list core versions
# Endpoint for listing all core dataset versions
endpoint = "list_core_versions"

def list_core_versions():
    url = f"{base_url}{endpoint}"
    response = requests.get(url)
    if response.status_code == 200:
        return response.json()
    else:
        print("Error fetching core versions:", response.status_code, response.text)
        return None

# Call the function
core_versions = list_core_versions()

# Print the retrieved core dataset versions
if core_versions is not None:
    print(core_versions)

# get list metadata version
endpoint_list = "list_metadata_versions"

def list_metadata_versions():
    url = f"{base_url}{endpoint_list}"
    response = requests.get(url)
    if response.status_code == 200:
        return response.json()
    else:
        print("Error fetching metadata versions:", response.status_code, response.text)
        return None

metadata_versions = list_metadata_versions()
metadata_versions

# get_core_dataset function
key = "MPDS2024a"
def get_core_dataset(api_key, key):
    url = f"{base_url}get_core?key={key}&api_key={api_key}"
    response = requests.get(url)
    if response.status_code == 200:
        return response.json()
    else:
        print("Error fetching core dataset:", response.status_code, response.text)
        return None

# retrieve the core dataset
core_dataset = get_core_dataset(api_key, key)
core_dataset

#get_metadata
version = "2024-1"
def get_metadata(api_key, party, date, version):
    keys = f"{party}_{date}"
    url = f"{base_url}metadata?api_key={api_key}&keys[]={keys}&version={version}"
    print(url)                                             
    response = requests.get(url) 
    if response.status_code == 200:
        return response.json()
    else:
        print("Error fetching metadata:", response.status_code, response.text)
        return None

if core_dataset:
        # The first entry is column headers, retrieve data rows
        headers = core_dataset[0]
        data_rows = core_dataset[1:]
        
        # Identify the index of the 'party' column
        party_index = headers.index('party')
        country_index = headers.index('country')

        # Filter to get entries for the specified Swiss parties
        swiss_party_ids = ['43110', '43120', '43320', '43420', '43520', '43531', '43810', '43811']
        swiss_parties = [row for row in data_rows if row[party_index] in swiss_party_ids]
        
        # Print the filtered entries for Swiss parties
        print(swiss_parties)


# get manifesto text
def get_texts_and_annotations(api_key, queryable_keys,version):
    url = f"{base_url}texts_and_annotations?api_key={api_key}&keys[]={queryable_keys}&version={version}"
    print(url)                                             
    response = requests.get(url) 
    if response.status_code == 200:
        return response.json()
    else:
        print("Error fetching metadata:", response.status_code, response.text)
        return None

queryable_keys = [f"{row[party_index]}_{row[headers.index('date')]}" for row in swiss_parties]
queryable_keys

# Initialize a list to store the data
data = []
# Loop over each queryable_key and fetch data
for key in queryable_keys:
    response = get_texts_and_annotations(api_key, key, version)
    if response and 'items' in response:
        for item in response['items']:
            party, date = key.split('_')
            for sub_item in item.get('items', []):
                text = sub_item.get('text', '')
                cmp_code = sub_item.get('cmp_code', '')
                eu_code = sub_item.get('eu_code', '')
                data.append({'party': party, 'date': date, 'text': text, 'cmp_code': cmp_code, 'eu_code': eu_code})


df = pd.DataFrame(data)


print(df)


df.to_csv('manifesto_texts.csv', index=False)

# get manifesto text english
def get_texts_and_annotations_en(api_key, queryable_keys,version):
    translation = "en" # returns translated text instead of original
    url = f"{base_url}texts_and_annotations?api_key={api_key}&keys[]={queryable_keys}&version={version}&translation={translation}"
    print(url)                                             
    response = requests.get(url) 
    if response.status_code == 200:
        return response.json()
    else:
        print("Error fetching metadata:", response.status_code, response.text)
        return None

# Repeat same steps but with translated data
#  Initialize a list to store the data
data_en = []

# Loop over each queryable_key and fetch data
for key in queryable_keys:
    response = get_texts_and_annotations_en(api_key, key, version)
    if response and 'items' in response:
        for item in response['items']:
            if not item.get('items'):  # Skip if 'items' is empty
                continue
            party, date = key.split('_')
            for sub_item in item.get('items', []):
                text = sub_item.get('text', '')
                cmp_code = sub_item.get('cmp_code', '')
                eu_code = sub_item.get('eu_code', '')
                data_en.append({'party': party, 'date': date, 'text': text, 'cmp_code': cmp_code, 'eu_code': eu_code})

# Create a DataFrame
df_en = pd.DataFrame(data_en)

# Print
print(df_en)

# Save the DataFrame to a CSV file
df_en.to_csv('./data/manifesto_texts_en.csv', index=False)

# Putting back the manifesto texts back together
df_en_combined = df_en.groupby(['party', 'date'])['text'].apply(' '.join).reset_index()
print(df_en_combined)

# combined DataFrame to a CSV file
df_en_combined.to_csv('./data/manifesto_combined_texts_en.csv', index=False)
