==> 4-annotation-model-1-chunk_texts.ipynb <==
# This script contains a function that chunks the text into sentences for unannotated manifestos from 1963 to 1995
import nltk
nltk.download('punkt')
from nltk.tokenize import sent_tokenize

import numpy as np
import pandas as pd


df_orig = pd.read_csv("./data/manifesto_texts.csv")[["party", "date", "text"]]
df = df_orig[df_orig["date"] < 199910]
df

all_manifestos = df["text"].to_list()

# Chunk all manifestos into sentences keeping date and party
all_dfs = []
for i, manifesto in enumerate(all_manifestos):
    sentences = sent_tokenize(manifesto)

    # Postprocessing (e.g., remove enumerations)
    sentences = [sentence for sentence in sentences if len(sentence) > 7]

    chunked_manifesto_df = pd.DataFrame({"party": np.nan, "date": np.nan, "text": sentences})
    chunked_manifesto_df["party"] = df.iloc[i, 0]
    chunked_manifesto_df["date"] = df.iloc[i, 1]
    all_dfs.append(chunked_manifesto_df)

dft = pd.concat(all_dfs)
dft

dft["len"] = dft["text"].str.len()
dft.sort_values("len")

dft[dft["len"] == 8]
# => Could do some more postprocessing


dft = dft.drop("len", axis=1)


dft.to_csv("./data/manifesto_texts_pre_99.csv", index=False)
