==> 6-annotation-model-3-check.ipynb <==
# This script serves to compare the Manifestoberta classifier with already annotated text
# It first chunks the text into sentences
# Second, it adds a topic label 

import nltk
nltk.download('punkt')
from nltk.tokenize import sent_tokenize

import numpy as np
import pandas as pd


df = pd.read_csv("./data/manifesto_combined_texts_en.csv")
df

df = df.sample(1)
df

manifesto_text = df["text"].tolist()
manifesto_text = str(manifesto_text)

# Chunk manifesto text into sentences
sentences = sent_tokenize(manifesto_text)

# Postprocessing (e.g., remove enumerations)
sentences = [sentence for sentence in sentences if len(sentence) > 7]

#This needs to be changed based on previous code chunk
party = "43420"
date = "201510"

chunked_manifesto_df = pd.DataFrame({"party": party, "date": date, "text": sentences})
chunked_manifesto_df

# Now apply the topic classifier 
import torch
from transformers import AutoTokenizer, AutoModelForSequenceClassification

# Check whether GPU available
device = "cuda:0" if torch.cuda.is_available() else "cpu"
device

# Classification model manifestoberta
model = AutoModelForSequenceClassification.from_pretrained("manifesto-project/manifestoberta-xlm-roberta-56policy-topics-sentence-2023-1-1")
tokenizer = AutoTokenizer.from_pretrained("xlm-roberta-large")
model = model.to(device)

# This function classifies sentences based on the 56 topics of the Manifesto Project
def classify_sentence(sentence):
    inputs = tokenizer(
        sentence,
        return_tensors="pt",
        # max_length=200,
        padding="max_length",
        truncation=True
    )

    # Move to GPU if available 
    inputs = inputs.to(device)

    logits = model(**inputs).logits
    predicted_class = model.config.id2label[logits.argmax().item()]
    return predicted_class

classified_manifesto = chunked_manifesto_df
classified_manifesto.loc[:, "topic"] = classified_manifesto["text"].apply(classify_sentence)
classified_manifesto

# Function to map topic to main_topic
def map_to_main_topic(topic):
    if str(topic).startswith('1'):
        return "external relations"
    elif str(topic).startswith('2'):
        return "freedom and democracy"
    elif str(topic).startswith('3'):
        return "political system"
    elif str(topic).startswith('4'):
        return "economy"
    elif str(topic).startswith('5'):
        return "welfare and quality of life"
    elif str(topic).startswith('6'):
        return "fabric of society"
    elif str(topic).startswith('7'):
        return "social groups"
    else:
        return "other" 

# Add a new column 'main_topic' based on the 'topic'
classified_manifesto['main_topic'] = classified_manifesto['topic'].apply(map_to_main_topic)
classified_manifesto

# Split 'topic' into 'cmp_code'
manifesto_text["cmp_code"] = manifesto_text["topic"].str.split(" - ", n=1, expand=True).loc[:,0]
manifesto_text

# Compute the shares of topics in order to compare with already annotated manifestos
counts = classified_manifesto["main_topic"].value_counts()
total = len(classified_manifesto["main_topic"])
share = (counts / total) * 100
share

# Compute share from annotated manifesto
# Load dataset from siebert sentiment analysis which contains manifesto from 1999 to 2019
annotated_manifestos_1999_2019= pd.read_pickle("./data/df_main_sentiment_1999-2019.pkl")
annotated_manifestos_1999_2019

#Select manifesto
selected_manifesto = annotated_manifestos_1999_2019[(annotated_manifestos_1999_2019["party"] == 43420) & (annotated_manifestos_1999_2019["date"] == 201510)]
selected_manifesto

# Compute the shares of topics in order to compare with already annotated manifestos
counts_selected = selected_manifesto["main_topic"].value_counts()
total_selected = len(selected_manifesto["main_topic"])
share_selected = (counts_selected / total_selected) * 100

# Compare shares of topic between two methods
print(f"Share of main topics using Manifestoberta classifier:",share)
print(f"Share of main topics of annotated manifestos;", share_selected)
