==> 7-sentiment-2-xlm-twitter-political.ipynb <==
import pandas as pd
import numpy as np
import torch
from transformers import AutoModelForSequenceClassification
from transformers import TFAutoModelForSequenceClassification
from transformers import AutoTokenizer
from scipy.special import softmax

df = pd.read_csv("./data/classified_manifestos_pre99.csv")
df.head()

# Split the topic into the cmp_code
df["cmp_code"] = df["topic"].str.split(" - ", n=1, expand=True).loc[:,0]
df 

# Add dictionary for mapping party IDs to party names
party_name_mapping = {
    43110: "GPS",
    43120: "GLP",
    43320: "SPS",
    43420: "FDP",
    43520: "CVP",
    43810: "SVP",
    43811: "BDP"
}

# Map the party IDs to the new 'party_name' column
df['party_name'] = df['party'].map(party_name_mapping)

df

# Function to map cmp_code to main_topic
def map_to_main_topic(cmp_code):
    if str(cmp_code).startswith('1'):
        return "external relations"
    elif str(cmp_code).startswith('2'):
        return "freedom and democracy"
    elif str(cmp_code).startswith('3'):
        return "political system"
    elif str(cmp_code).startswith('4'):
        return "economy"
    elif str(cmp_code).startswith('5'):
        return "welfare and quality of life"
    elif str(cmp_code).startswith('6'):
        return "fabric of society"
    elif str(cmp_code).startswith('7'):
        return "social groups"
    else:
        return "other"  # Optional: For any other cases

# Add a new column 'main_topic' based on the 'cmp_code'
df['main_topic'] = df['cmp_code'].apply(map_to_main_topic)

df.head()

# Save dataset for pre-1999 manifestos
df.to_pickle("./data/df_pre1999.pkl")

# Load dataset for 1999 and post manifestos
df_post1999 = pd.read_pickle("./data/df_all.pkl")
df_post1999

# Unify date formats and drop columns
df_convert = pd.read_pickle("./data/df_pre1999.pkl")
df_convert["date"]= pd.to_datetime(df_convert["date"], format="%Y%m")
drop_col = ["topic"]
df_pre1999 = df_convert.drop(drop_col, axis=1)
df_pre1999

# Drop unnecessary columns in post 1999 dataset
drop_col = ["eu_code","sentiment", "score","year"]
df_post1999 = df_post1999.drop(drop_col, axis=1)
df_post1999

# Combine the two datasets
dfs = [df_pre1999, df_post1999]
dfs = pd.concat(dfs)
dfs.shape

# Save dataframe for the sentiment analysis
dfs.to_csv("./data/manifesto_texts_topics_1963_2019.csv")

# Load dataset
df_orig = pd.read_csv("/content/drive/MyDrive/dlss24/dlss24/final/manifesto_texts_topics_1963_2019.csv")
df_orig

# Load model
MODEL = f"cardiffnlp/xlm-twitter-politics-sentiment"
model = AutoModelForSequenceClassification.from_pretrained(MODEL)
tokenizer = AutoTokenizer.from_pretrained(MODEL)
model = model.to(device)

device = "cuda:0" if torch.cuda.is_available() else "cpu"
device

# Function for sentiment
def sentiment_analyzer_sentences (text):
  encoded_input = tokenizer(text, return_tensors='pt', truncation=True, padding='max_length', max_length=512)
  # Move to GPU
  encoded_input = encoded_input.to(device)

  output = model(**encoded_input)
  scores = output[0][0].cpu().detach().numpy()
  scores = softmax(scores)
  # assigning sentiment label based on highest score index
  sentiment_index = scores.argmax()
  sentiments = ["negative", "neutral", "positive"]
  sentiment = sentiments[sentiment_index]
  return sentiment

# Apply sentiment function to text 
df = df_orig.copy()
df.loc[:,"sentiment"] = df["text"].apply(sentiment_analyzer_sentences)

# Saved dataframe
df.to_csv("./data/manifesto-texts-sentiment-1963-2019.csv")
